Intial import.
crawl.py: A QtWebKit based web crawler that is designed to click on everything, even things in iframes, for a list of links and to a certain depth, with the goal of investigating distribution of malware via advertising networks. While it can crawl through all links on a set of pages, it includes optimizations to reduce crawling time. Fully headless, but based on a real web browser (WebKit) so it loads all Javascript correctly.
get-links.py: A QtWebKit based link scraper for a single page.